加微信小程序之前
加微信小程序之前

+++ catcate-cutomer/src/main/java/com/zrrd/catcatecutomer/service/impl/ProductsServiceImpl.java
@@ -155,7 +155,7 @@ public class ProductsServiceImpl extends ServiceImpl<ProductsMapper, Products> i
 
             if (!reviews.isEmpty()) {
                 double avgRating = reviews.stream()
-                        .mapToInt(Reviews::getRating)
+                        .mapToDouble(Reviews::getRating)
                         .average()
                         .orElse(0.0);
